requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-groq==0.1.0
langchain-core==0.1.0
groq==0.4.0
pdfplumber==0.10.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
supabase==2.3.0
python-dotenv==1.0.0
validators==0.22.0
python-dateutil==2.8.2
reportlab==4.0.7
pandas==2.1.0
numpy==1.24.3

